refactor(engine): extract websocket health watchdog
