+++ var/agents/configuration/environment_variables.json
@@ -906,7 +906,7 @@
       ],
       "code_references": [
         {
-          "line": 887,
+          "line": 888,
           "path": "src/gpt_trader/features/live_trade/engines/strategy.py",
           "reader": "os.getenv"
         }
