+++ var/agents/logging/event_catalog.json
@@ -2554,7 +2554,8 @@
       ],
       "occurrence_count": 7,
       "source_files": [
-        "features/live_trade/engines/strategy.py"
+        "features/live_trade/engines/strategy.py",
+        "features/live_trade/engines/ws_health_monitor.py"
       ],
       "statuses": null
     }
